never give up

+++ python-if_else_loops_functions/9-print_last_digit.py
@@ -1,4 +1,4 @@
 #!/usr/bin/python3
 def print_last_digit(number):
-    number = int(number)
+    number = float(number)
     return number % 10
